Trying to create a custom evaluator for Functional programming
