+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/MockingEvaluator.java
@@ -5,13 +5,9 @@ import com.github.javaparser.ast.body.CallableDeclaration;
 import com.github.javaparser.ast.body.MethodDeclaration;
 import com.github.javaparser.ast.expr.Expression;
 import com.github.javaparser.ast.type.Type;
-import sa.com.cloudsolutions.antikythera.exception.EvaluatorException;
 import sa.com.cloudsolutions.antikythera.parser.AbstractCompiler;
-import sa.com.cloudsolutions.antikythera.parser.Callable;
-import sa.com.cloudsolutions.antikythera.parser.MCEWrapper;
 
 import java.util.Map;
-import java.util.Optional;
 
 public class MockingEvaluator extends AbstractEvaluator implements ExpressionEvaluator {
 
@@ -22,15 +18,6 @@ public class MockingEvaluator extends AbstractEvaluator implements ExpressionEva
     @Override
     public Variable evaluateExpression(Expression expr) throws ReflectiveOperationException {
         return null;
-    }
-
-    @Override
-    public CompilationUnit getCompilationUnit() {
-        return null;
-    }
-
-    @Override
-    public void setCompilationUnit(CompilationUnit compilationUnit) {
     }
 
     @Override
